Split out different parts of the preload API, tests still work
