+++ packages/preload/src/fdc3-1.2/index.ts
@@ -1,8 +1,11 @@
-import { connect, createAPI } from './api';
-import { contextBridge } from 'electron';
-connect();
+import { sendMessage } from '../lib/lib';
+import { createAPI } from './api';
+import { connect } from './connect';
+import { contextBridge, ipcRenderer } from 'electron';
+
+connect(ipcRenderer);
 //listen();
 //document.addEventListener('DOMContentLoaded', createAPI);
-const DesktopAgent = createAPI();
+const desktopAgent = createAPI(sendMessage, ipcRenderer);
 /* expose the fdc3 api across the context isolation divide...*/
-contextBridge.exposeInMainWorld('fdc3', DesktopAgent);
+contextBridge.exposeInMainWorld('fdc3', desktopAgent);
